Use new show custom input method information to display option change
Instead of using system notification, which we do not have a long,
descriptive string, use the input popup to display such information
instead.

This can be easily achieved, by dispatch "option" notification in main
thread to get the changed options. Since key event always happens in the
main thread, we can guarantee that there is only one RimeState handling
the key event. After that, try to aggregate based on the
RimeOptionAction so the 3+ state option will not display multiple
strings.

Fix #111